Refactor loading into loaders module
